LPS-61969 move control menu entry to its owner
